After rearrangement of classes
