+++ src/tricera/concurrency/CCAstTypeAnnotator.scala
@@ -47,63 +47,6 @@ class CCAstTypeAnnotator {
   }
 }
 
-/*
-   The CCAstDeclaration represents a declaration in the AST.
-   It can be used to construct various kinds of nodes when
-   manipulating the tree.
-*/
-object CCAstDeclaration {
-  private val copyAst = new CCAstCopyVisitor
-  private val getName = new CCAstGetNameVistor
-  private val getDeclarator = new CCAstGetDeclaratorVistor
-
-  def apply(d: ListDeclaration_specifier, i: Init_declarator, e: ListExtra_specifier): CCAstDeclaration = {
-    new CCAstDeclaration(d, i, e)
-  }
-
-  def apply(d: ListDeclaration_specifier, i: Init_declarator): CCAstDeclaration = {
-    new CCAstDeclaration(d, i, new ListExtra_specifier)
-  }
-} 
-
-class CCAstDeclaration(d: ListDeclaration_specifier, i: Init_declarator, e: ListExtra_specifier) {
-  import CCAstDeclaration._
-  private val declarationSpecifiers: ListDeclaration_specifier = d
-  private val initDeclarator: Init_declarator = i
-  private val extraSpecifiers: ListExtra_specifier = e
-
-  def getId(): String = {
-    initDeclarator.accept(getName, ())
-  }
-
-  def toAfunc(body: Compound_stm): Afunc = {
-    new Afunc(
-      new NewFunc(
-        copyAst(declarationSpecifiers),
-        initDeclarator.accept(getDeclarator, ()),
-        body))
-  }
-
-  def toDeclarators(): Declarators = {
-    val initDecls = new ListInit_declarator
-    initDecls.add(initDeclarator.accept(copyAst, ()))
-    new Declarators(
-      copyAst(declarationSpecifiers),
-      initDecls,
-      copyAst(e))
-  }
-
-  def toEvarWithType(): EvarWithType = {
-    val decSpecs = copyAst(declarationSpecifiers)
-    val initDec = initDeclarator.accept(copyAst, ())
-    new EvarWithType(initDeclarator.accept(getName, ()), decSpecs, initDec);
-  }
-
-  def toGlobal(): Global = {
-    new Global(toDeclarators())
-  }
-}
-
 class CCAstTypeAnnotationData {
   private val globalVariables = MHashMap[String, CCAstDeclaration]()
 
@@ -114,132 +57,6 @@ class CCAstTypeAnnotationData {
   def get(name: String): Option[CCAstDeclaration] = {
     globalVariables.get(name)
   }
-}
-
-/*
-  Vistor class to extract a name from a declaration or definition.
-*/
-class CCAstGetNameVistor extends AbstractVisitor[String, Unit] {
-    /* External_declaration */
-    override def visit(ext: Afunc, arg: Unit): String = { ext.function_def_.accept(this, arg) }
-    override def visit(ext: Athread, arg: Unit): String = { ext.thread_def_.accept(this, arg) }
-    override def visit(ext: Global, arg: Unit): String = { ext.dec_.accept(this, arg) }
-    override def visit(ext: Chan, arg: Unit): String = { ext.chan_def_.accept(this, arg) }
-
-    /* Function_def */
-    override def visit(defn: AnnotatedFunc, arg: Unit): String = { defn.declarator_.accept(this, arg); }
-    override def visit(defn: NewFuncInt, arg: Unit): String = { defn.declarator_.accept(this, arg); }
-    override def visit(defn: NewFunc, arg: Unit): String = { defn.declarator_.accept(this, arg); }
-
-    /* Thread_def */
-    override def visit(trd: SingleThread, arg: Unit): String = { trd.cident_; }
-    override def visit(trd: ParaThread, arg: Unit): String = { trd.cident_2; }
-
-    /* Dec */
-    override def visit(decs: Declarators, arg: Unit): String = {
-      decs.listinit_declarator_.asScala.map(d => d.accept(this, arg)).filter(n => n != "").head
-    }
-    
-    /* Init_declarator */
-    override def visit(dec: OnlyDecl, arg: Unit): String = { dec.declarator_.accept(this, arg); }
-    override def visit(dec: InitDecl, arg: Unit): String = { dec.declarator_.accept(this, arg); }
-    override def visit(dec: HintDecl, arg: Unit): String = { dec.declarator_.accept(this, arg); }
-    override def visit(dec: HintInitDecl, arg: Unit): String = { dec.declarator_.accept(this, arg); }
-
-    /* Declarator */
-    override def visit(dec: BeginPointer, arg: Unit): String = { return dec.direct_declarator_.accept(this, arg); }
-    override def visit(dec: NoPointer, arg: Unit): String = { return dec.direct_declarator_.accept(this, arg); }
-
-    /* Direct_declarator */
-    override def visit(dec: Name, arg: Unit): String = { return dec.cident_; }
-    override def visit(dec: ParenDecl, arg: Unit): String = { return dec.declarator_.accept(this, arg); }
-    override def visit(dec: InitArray, arg: Unit): String = { return dec.direct_declarator_.accept(this, arg); }
-    override def visit(dec: Incomplete, arg: Unit): String = { return dec.direct_declarator_.accept(this, arg); }
-    override def visit(dec: MathArray, arg: Unit): String = { return dec.direct_declarator_.accept(this, arg); }
-    override def visit(dec: NewFuncDec, arg: Unit): String = { return dec.direct_declarator_.accept(this, arg); }
-    override def visit(dec: OldFuncDec, arg: Unit): String = { return dec.direct_declarator_.accept(this, arg); }
-
-    /* Parameter_declaration */
-    override def visit(param: TypeAndParam, arg: Unit): String = { param.declarator_.accept(this, arg) }
-    override def visit(param: TypeHintAndParam, arg: Unit): String = { param.declarator_.accept(this, arg) }
-
-    /* Exp */
-    override def visit(exp: Earray, arg: Unit) : String = { exp.exp_1.accept(this, ()) }
-    override def visit(exp: Efunk, arg: Unit): String = { exp.exp_.accept(this, ()) }
-    override def visit(exp: Efunkpar, arg: Unit): String = { exp.exp_.accept(this, ()) }
-    override def visit(exp: Evar, arg: Unit): String = { exp.cident_ }
-    override def visit(exp: EvarWithType, arg: Unit): String = { exp.cident_ }
-}
-
-/*
-  Vistor to copy an AST.
-*/
-class CCAstCopyVisitor extends ComposVisitor[Unit] {
-  def apply(specifiers: ListDeclaration_specifier): ListDeclaration_specifier = {
-    val decSpecifiers = new ListDeclaration_specifier
-    for (d <- specifiers.asScala)
-    {
-      decSpecifiers.add(d.accept(this, ()));
-    }
-    decSpecifiers
-  }
-
-  def apply(specifiers: ListExtra_specifier): ListExtra_specifier = {
-    val extraSpecifiers = new ListExtra_specifier
-    for (s <- specifiers.asScala)
-    {
-      extraSpecifiers.add(s.accept(this, ()));
-    }
-    extraSpecifiers
-  }
-
-  def apply(expressions: ListExp) = {
-    val expsCopy = new ListExp
-    for (x <- expressions.asScala) {
-      expsCopy.add(x.accept(this, ()))
-    }
-    expsCopy
-  }
-
-  def apply(params: ListParameter_declaration) = {
-    val paramsCopy = new ListParameter_declaration
-    for (p <- params.asScala) {
-      paramsCopy.add(p.accept(this, ()))
-    }
-    paramsCopy
-  }
-}
-
-/*
-  Vistor to extract a function declaration from a function definition.
-*/
-class CCAstGetFunctionDeclarationVistor extends AbstractVisitor[(ListDeclaration_specifier, Init_declarator), Unit] {
-  val copyAst = new CCAstCopyVisitor
-  /* Function_def */
-  override def visit(defn: AnnotatedFunc, arg: Unit) = { 
-    (copyAst(defn.listdeclaration_specifier_), new OnlyDecl(defn.declarator_.accept(copyAst, arg)));
-  }
-  override def visit(defn: NewFuncInt, arg: Unit) = {
-    val declarationSpecifiers = new ListDeclaration_specifier
-    declarationSpecifiers.add(new Type(new Tint))
-    (declarationSpecifiers, new OnlyDecl(defn.declarator_.accept(copyAst, arg))); 
-  }
-  override def visit(defn: NewFunc, arg: Unit) = { 
-    (copyAst(defn.listdeclaration_specifier_), new OnlyDecl(defn.declarator_.accept(copyAst, arg)));
-  }
-}
-
-/*
-  Vistor to extract the "Declarator" part from an Init_declaration
-*/
-class CCAstGetDeclaratorVistor extends AbstractVisitor[Declarator, Unit] {
-  private val copyAst = new CCAstCopyVisitor
-
-  /* Init_declarator */
-  override def visit(dec: OnlyDecl, arg: Unit): Declarator = { dec.declarator_.accept(copyAst, arg); }
-  override def visit(dec: InitDecl, arg: Unit): Declarator = { dec.declarator_.accept(copyAst, arg); }
-  override def visit(dec: HintDecl, arg: Unit): Declarator = { dec.declarator_.accept(copyAst, arg); }
-  override def visit(dec: HintInitDecl, arg: Unit): Declarator = { dec.declarator_.accept(copyAst, arg); }
 }
 
 /*
